backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
supabase==2.3.1
numpy==1.26.4
torch==2.2.1
torchvision==0.17.1
nibabel==5.2.1
pydicom==2.4.4
scikit-image==0.22.0
scikit-learn==1.4.0
pandas==2.2.0
matplotlib==3.8.3
seaborn==0.13.2
starlette==0.36.3
httpx==0.24.1
websockets==11.0.3
tensorflow>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0
